Enforces only deleting your own comments.

+++ components/Comments/CommentContent.tsx
@@ -51,30 +51,36 @@ const CommentContent: React.FC<CommentContentProps> = ({
 	};
 
 	const deleteHandler: MouseEventHandler<HTMLButtonElement> = async () => {
-		if (confirm("Are you sure you want to delete this comment?")) {
-			try {
-				if (isSecondLevelComment) {
-					if (originalComment && originalComment.replies) {
-						const deletedComment = {
-							...originalComment,
-							replies: originalComment.replies.filter(
-								(reply) => reply._id !== currComment._id
-							),
-						};
-						postUpdatedResourceToDb(deletedComment, originalComment._id);
-					}
-				} else {
-					await fetch(`/api/comments/${currComment._id}`, {
-						method: "DELETE",
-						headers: {
-							"Content-Type": "application/json",
-						},
-					});
-				}
+		if (dbUser) {
+			if (dbUser.userId === currComment.authorId) {
+				if (confirm("Are you sure you want to delete this comment?")) {
+					try {
+						if (isSecondLevelComment) {
+							if (originalComment && originalComment.replies) {
+								const deletedComment = {
+									...originalComment,
+									replies: originalComment.replies.filter(
+										(reply) => reply._id !== currComment._id
+									),
+								};
+								postUpdatedResourceToDb(deletedComment, originalComment._id);
+							}
+						} else {
+							await fetch(`/api/comments/${currComment._id}`, {
+								method: "DELETE",
+								headers: {
+									"Content-Type": "application/json",
+								},
+							});
+						}
 
-				await scrollToSamePosition();
-			} catch (error) {
-				console.error(error);
+						await scrollToSamePosition();
+					} catch (error) {
+						console.error(error);
+					}
+				}
+			} else {
+				alert("This ain't your comment to delete!");
 			}
 		}
 		setOptionsVisible(false);
